requirements.txt
fastapi[all]
pyyaml
pytest
motor~=3.0.0
websockets~=10.3